pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-forge"
version = "1.0.0"
description = "AI-Powered One-Shot App Builder with Multi-Agent Orchestration"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Nexus Forge Team", email = "team@nexusforge.ai"}
]
maintainers = [
    {name = "Nexus Forge Team", email = "team@nexusforge.ai"}
]
keywords = [
    "ai", "app-builder", "multi-agent", "orchestration", 
    "gemini", "google-cloud", "fastapi", "react"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-cloud-monitoring>=2.16.0",
    "google-cloud-logging>=3.8.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-run>=0.10.0",
    "google-cloud-resource-manager>=1.10.0",
    "google-auth>=2.25.0",
    "vertexai>=1.38.0",
    "google-generativeai>=0.8.0",
    "PyGithub>=1.59.0",
    "Pillow>=10.1.0",
    "aiofiles>=23.2.0",
    "stripe>=7.8.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "asyncio-mqtt>=0.14.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0"
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0"
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocs-swagger-ui-tag>=0.6.0"
]

[project.urls]
Homepage = "https://nexusforge.ai"
Documentation = "https://docs.nexusforge.ai"
Repository = "https://github.com/nexus-forge/nexus-forge"
Issues = "https://github.com/nexus-forge/nexus-forge/issues"
Changelog = "https://github.com/nexus-forge/nexus-forge/blob/main/CHANGELOG.md"

[project.scripts]
nexus-forge = "nexus_forge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexus_forge*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["nexus_forge"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.cloud.*",
    "stripe.*",
    "redis.*",
    "celery.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short --strict-markers --disable-warnings"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "e2e: End-to-end tests",
    "slow: Slow tests"
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["nexus_forge"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]